+++ money-akka/src/main/scala/com/comcast/money/akka/stream/StreamTracingDSL.scala
@@ -454,43 +454,3 @@ object StreamTracingDSL {
     (output, traceContext)
   }
 }
-
-/**
- * Async Tracing for Unordered Flows
- */
-
-object AsyncUnorderedFlowTracing {
-
-  implicit class TracedFlowOps[In: ClassTag, Out: ClassTag](flow: Flow[In, Out, _])(implicit executionContext: ExecutionContext, fskc: FlowSpanKeyCreator[In] = DefaultFlowSpanKeyCreator[In]) {
-    type TracedIn = (In, TraceContext)
-    type TracedOut = (Out, TraceContext)
-
-    /**
-     * Applies a traced version of [[Flow.mapAsyncUnordered]] to a generic [[Flow]].
-     * Guarantees that elements are executed out of order and are attached to the SpanContext they entered with.
-     *
-     * NB: It is necessary to use flatMap as map has been found to not be guaranteed to close the Span
-     *
-     * Example:
-     *
-     * Flow[String].tracedMapAsyncUnordered(3)(stringToFuture)
-     *
-     * @param parallelism the number of concurrent futures
-     * @param f           the function to apply
-     * @return Flow[TracedIn, TracedOut, _]
-     */
-
-    def tracedMapAsyncUnordered(parallelism: Int)(f: In => Future[Out]): Flow[TracedIn, TracedOut, _] =
-      Flow[TracedIn].mapAsyncUnordered[TracedOut](parallelism) {
-        (tuple: TracedIn) =>
-          val (in, traceContext) = tuple
-          traceContext.tracer.startSpan(fskc.flowToKey(flow))
-          f(in) map {
-            out =>
-              traceContext.tracer.stopSpan()
-              (out, traceContext)
-          }
-      }
-  }
-
-}
